refactor: move files for layer packaging
