+++ wasin/src/main/java/com/wasin/backend/controller/AlertController.java
@@ -1,6 +1,6 @@
 package com.wasin.backend.controller;
 
-import com.wasin.backend.domain.alert.service.AlertService;
+import com.wasin.backend.service.AlertService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/api/notifications")
+@RequestMapping("/api/alert")
 public class AlertController {
 
     private final AlertService alertService;
